iocBoot: override TOP variable to point to "../../"
This is needed as the compilation path is
different that the execution path.

Generally, we install our IOCs in
"/opt/epics/startup/ioc/${ioc_name}".
So, the TOP variable from envPaths will
point to the wrong location.

Most of the time, things will work out
just fine, as the same files will be in both
places, but autosave locations, for instance,
will be wrong.