Add FPGAClkMAF and FPGAClkCalc PVs.
FPGAClkMAF is a 5-sample moving average of the
Freq-Mon, which is the FPGA frequency measured
by hardware.

The PV FPGAClk-Cte shows the FPGA clock frequency,
which comes from FPGAClkMAF or can be defined by
user.

FPGAClkCalc sets FPGAClk-Cte with the FPGAClkMAF
value if TimestampSrc-Sel is not IDLE, otherwise
it holds the FPGAClk-Cte value set by user.