timingApp: implement new version string.
The 36 character limit in CALC made it necessary to define 3
intermediary records (even if using the "$P" shorthand for PRINTF).

Since the firmware version logic is the same for all hardware types, we
can have a single .db file for them. We also change it to use Fw instead
of Frm to refer to "firmware".

Since the records are only processed at startup (using callbacks), we
can omit the in/out commands outside of the @init section.