timingApp: add RxLocked.
And add the sequencer implementation for its logic.

The idea is that a single change in FramePos in a "long" time frame
(such as 5s, considering the 1s SCAN on the record) can be due to
long-term drifts; more than one change in the same time frame can be
attributed the the downlink's RX clock not being properly locked (the
main situation where this can happen is with the AFC Timing).