timingApp: process RxEnbl-SP during startup.
Otherwise, other proto commands will use an uninitialized RxEnblHw-SP
value during startup, and RxEnbl for those outputs will be disabled.
This doesn't change the fact that RxEnblHw-SP will be used uninitialized
(so RxEnbl will be disabled for all outputs momentarily), but it makes
it so that, at the end of iocInit, the final value written to hardware
will be correct. This is specially relevant for these PVs because writes
to them are disabled by the interlock status, so forcing them to process
to fix this erroneous condition would require momentarily disabling the
interlock.