event_log: clean up event log reset command.
The reset command was using the sseq record unnecessarily, because
"Wait" values for WAITx fields only work when using CA links, which
can't be used here, because they don't cause the compress records to be
processed.

We are enquiring about the reliability of the reset mechanism used here
in tech-talk [1]. We could use a single write to RES, without PP, but
that would mean the interface is only updated when the next event
arrives into the log, instead of having immediate effect, which doesn't
make for the best user experience.

Since we added processing to the LOGSOFTCNT link, it was necessary to
write -1 into it to guarantee the final value is 0.

[1] https://epics.anl.gov/tech-talk/2024/msg01373.php