event_log: add EventTable NTTable.
Instead of requiring clients to combine multiple waveforms themselves,
provide a table with timestamp and event information.

It was necessary to add a top level trigger mapping to rstSoftBuff,
otherwise, when resetting the buffers, the table would have one element
from each buffer, due to their spurious processing; so we needed to
fetch the values again after the buffer reset was completed.